+++ apps/web/components/CardWithFeaturedItems/CardWithFeaturedItems.css.ts
@@ -1,19 +1,12 @@
-import { globalStyle, style } from '@vanilla-extract/css'
+import { style } from '@vanilla-extract/css'
 
-import { theme, themeUtils } from '@island.is/island-ui/theme'
+import { themeUtils } from '@island.is/island-ui/theme'
 
 export const container = style({
   height: 260,
   minWidth: 0,
   minHeight: 0,
   overflow: 'hidden',
-})
-
-export const purpleTags = style({})
-
-globalStyle(`${purpleTags} a:hover, ${purpleTags} button:hover`, {
-  backgroundColor: theme.color.purple400,
-  color: theme.color.white,
 })
 
 export const image = style({
