+++ apps/web/screens/Organizations/Organizations.css.ts
@@ -4,6 +4,8 @@ import { theme } from '@island.is/island-ui/theme'
 
 const mobileMediaQuery = `screen and (max-width: ${theme.breakpoints.md - 1}px)`
 
+export { heading, description, listContainer } from '../shared/listingPageStyles.css'
+
 export const filterBar = style({
   display: 'flex',
   alignItems: 'center',
@@ -41,23 +43,4 @@ export const radioGroup = style({
 export const filterButton = style({
   flexShrink: 0,
   flexGrow: 0,
-})
-
-export const heading = style({
-  fontSize: 32,
-  paddingTop: 16,
-})
-
-export const description = style({
-  fontSize: 18,
-  maxWidth: 774,
-})
-
-export const listContainer = style({
-  '@media': {
-    [mobileMediaQuery]: {
-      paddingLeft: 24,
-      paddingRight: 24,
-    },
-  },
 })
